Keep calm and carry on
Just don't do all the things when all the things are not possible. I
mean why panic!()? What do we gain, as a society?

When it doesn't make sense to propagate a trait impl item to the enum,
we just skip it and drop the trait impl from the enum, instead providing
whatever forwarding methods we can in an ordinary impl block.